+++ bliss/common/plot.py
@@ -169,6 +169,7 @@ get_flint = flint_proxy.get_flint
 check_flint = flint_proxy.check_flint
 attach_flint = flint_proxy.attach_flint
 reset_flint = flint_proxy.reset_flint
+restart_flint = flint_proxy.restart_flint
 
 
 def _create_plot(